NewsReporterAIParallel/requirements.txt
crewai
openai
python-dotenv
requests
crewai-tools
streamlit
watchdog